requirements.txt
# Discord Bot Libraries
discord.py>=2.0.0
nextcord>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# Music/Media
wavelink>=2.0.0

# Trading/Binance
python-binance>=1.0.0

# Data Processing
numpy>=1.24.0

# Lyrics
lyricsgenius>=3.7.5